Split remaining commands & cleaned up
