+++ packages/core/src/blocks/ParagraphBlockContent/ParagraphBlockContent.ts
@@ -1,4 +1,4 @@
-import { updateBlockCommand } from "../../api/blockManipulation/commands/updateBlock.js";
+import { updateBlockCommand } from "../../api/blockManipulation/commands/updateBlock/updateBlock.js";
 import { getCurrentBlockContentType } from "../../api/getCurrentBlockContentType.js";
 import {
   createBlockSpecFromStronglyTypedTiptapNode,
