+++ packages/core/src/extensions/KeyboardShortcuts/KeyboardShortcutsExtension.ts
@@ -1,9 +1,9 @@
 import { Extension } from "@tiptap/core";
 
 import { TextSelection } from "prosemirror-state";
-import { mergeBlocksCommand } from "../../api/blockManipulation/commands/mergeBlocks.js";
-import { splitBlockCommand } from "../../api/blockManipulation/commands/splitBlock.js";
-import { updateBlockCommand } from "../../api/blockManipulation/commands/updateBlock.js";
+import { mergeBlocksCommand } from "../../api/blockManipulation/commands/mergeBlocks/mergeBlocks.js";
+import { splitBlockCommand } from "../../api/blockManipulation/commands/splitBlock/splitBlock.js";
+import { updateBlockCommand } from "../../api/blockManipulation/commands/updateBlock/updateBlock.js";
 import { getBlockInfoFromPos } from "../../api/getBlockInfoFromPos.js";
 import { BlockNoteEditor } from "../../editor/BlockNoteEditor.js";
